Add auto-increment feature to versioning script
The versioning extension now has `incrementPositionIfSnapshot` and
`skipIncrement` for controlling what and how to increment the version
number automatically for snapshot versions. See the javadocs in the
script for more information.

This is useful for auto-publishing releases, so snapshots are correctly
under e.g. the next minor release version rather than snapshots of
the most recent release.